requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
torch>=1.10.0
torchvision>=0.11.0
PyYAML>=6.0
tqdm>=4.64.0
pillow>=9.0.0
matplotlib>=3.5.0
scipy>=1.7.0
